+++ src/fn/fn-colors.js
@@ -1,4 +1,0 @@
-'use strict';
-
-module.exports = require('./fn-anonymous-tuple');
-module.exports.fnName = 'colors';